Добавлена 6 и 7 работа

+++ src/Controller/LessonController.php
@@ -10,7 +10,7 @@
  use Symfony\Component\HttpFoundation\Request;
  use Symfony\Component\HttpFoundation\Response;
  use Symfony\Component\Routing\Annotation\Route;
-
+ use Symfony\Component\Security\Core\Exception\AccessDeniedException;
  #[Route('/lessons')]
  class LessonController extends AbstractController
  {
@@ -23,8 +23,12 @@
      }
 
      #[Route('/{id}/edit', name: 'app_lesson_edit', methods: ['GET', 'POST'])]
+     
      public function edit(Request $request, Lesson $lesson, LessonRepository $lessonRepository): Response
      {
+        if (!$this->isGranted('ROLE_SUPER_ADMIN')) {
+            throw new AccessDeniedException('У вас нет доступа к этой операции');
+        }
          $form = $this->createForm(LessonType::class, $lesson);
          $form->handleRequest($request);
 
@@ -46,6 +50,9 @@
      #[Route('/{id}', name: 'app_lesson_delete', methods: ['POST'])]
      public function delete(Request $request, Lesson $lesson, LessonRepository $lessonRepository): Response
      {
+        if (!$this->isGranted('ROLE_SUPER_ADMIN')) {
+            throw new AccessDeniedException('У вас нет доступа к этой операции');
+        }
          $courseId = $lesson->getCourse()->getId();
          if ($this->isCsrfTokenValid('delete'.$lesson->getId(), $request->request->get('_token'))) {
              $lessonRepository->remove($lesson, true);
